--- components/Core/core.scss
@@ -1,11 +1,0 @@
-.search-main {
-  width: 60%;
-}
-
-.search-and-paginate {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  align-items: flex-start;
-}
