+++ pages/index.tsx
@@ -1,7 +1,4 @@
-import Link from 'next/link';
-import React from 'react';
 import HomePage from './HomePage';
-//import App from '../src/App';
 
 const index = () => {
   return (
