--- src/app.scss
@@ -1,25 +1,0 @@
-.app {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 0 2rem;
-  text-align: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  padding-top: 40px;
-  flex-wrap: wrap;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 10px;
-  padding: 10px;
-  max-width: 1280px;
-  margin: 0 auto;
-  margin-top: 30px;
-  border-radius: 10px;
-  background-color: #f5f5f5;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
