+++ tsconfig.app.json
@@ -3,11 +3,7 @@
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "target": "ES2020",
     "useDefineForClassFields": true,
-    "lib": [
-      "ES2020",
-      "DOM",
-      "DOM.Iterable"
-    ],
+    "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
     "skipLibCheck": true,
     "moduleResolution": "Bundler",
@@ -21,7 +17,7 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true,
-    "composite": true,
+    "composite": true
   },
   "include": [
     "src",
@@ -39,6 +35,10 @@
     "next-env.d.ts",
     "next.config.js",
     "next-i18next.config.js",
-    "tsconfig.json", "services", "service",
+    "tsconfig.json",
+    "services",
+    "service",
+    "tests",
+    "types"
   ]
 }
