+++ tsconfig.node.json
@@ -18,7 +18,7 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true,
 
-    "composite": true 
+    "composite": true
   },
   "include": ["vite.config.ts"]
 }
